photos uploading with description is working and logout is working

+++ server/controllers/post.js
@@ -1,11 +1,12 @@
 const Postmodel = require("../models/Postmodel")
+const cloudinary = require('cloudinary');
 
 
 // create a post
 // 
-exports.makeAPost = async (req,res)=>{
+exports.addProfilePicture = async (req,res)=>{
    req.body.userid = req.user._id
-
+  
    try {
     const newPost = new Postmodel(req.body)
     const savedPost = await newPost.save()
@@ -14,6 +15,138 @@ exports.makeAPost = async (req,res)=>{
     res.status(500).json(error.message)
   }
 }
+
+exports.addPosts = async (req, res, next) =>{
+  // images
+try {
+
+  
+  let imageArray = [];
+  let product;
+  if (!req.files) {
+   throw new Error("Images are not found");
+  }
+  // console.log("RESULT", req.files.photos);
+// console.log(req.files.photos.tempFilePath);
+
+  if (req.files) {
+    
+    for (let index = 0; index < req.files.photos.length; index++) {
+     
+      let result = await cloudinary.v2.uploader.upload(
+        req.files.photos[index].tempFilePath,
+        {
+          folder: "users",
+        }
+      );
+
+
+      // console.log("RESULT", result);
+      imageArray.push({
+        id: result.public_id,
+        secure_url: result.secure_url,
+      });
+    }
+
+   
+    //only one image
+    if(req.files.photos.length === undefined && req.files){
+    
+
+      try {
+       
+        let result = await cloudinary.v2.uploader.upload(
+          req.files.photos.tempFilePath,
+          {
+            folder: "users",
+          }
+        );
+        imageArray.push({
+          id: result.public_id,
+          secure_url: result.secure_url,
+        });
+      } catch (error) {
+        console.log("error",error);
+      return res.status(401).send(error)
+      }
+
+    }
+    
+    for (let index = 0; index < imageArray.length; index++) {
+      
+     product = await Postmodel.create({
+      userid:req.user.id,
+      img:{
+        id:imageArray[index].id,
+        secure_url:imageArray[index].secure_url 
+      }
+     })
+      // req.body.img = imageArray[index];
+      // req.body.userid = req.user.id;
+      // console.log("iam req.body ->",req.body);
+      //  product = await Postmodel.create(req.body);
+      // console.log("product",product);
+    
+     }
+
+       
+      res.status(200).json({
+        success: true,
+        product,
+      });  
+    
+  }
+} catch (error) {
+  console.log(error);
+  res.status(400).send(error.message)
+}
+};
+
+exports.descriptionPost = async (req,res)=>{
+   const {desc,imageID} = req.body
+   if(!desc){
+    return res.status(400).send("Description not found")
+   }
+   else if(imageID){
+    
+    try {
+      const image = await Postmodel.findById({_id:imageID})
+      console.log("image",image);
+         if(!image){
+          return res.status(404).send("the image is not uploaded")
+         }
+         image.description = desc
+         await image.save()
+       return  res.status(200).json({
+          success:true,
+  
+        })
+    } catch (error) {
+      res.status(401).send(error.message)
+    }
+    
+   }
+
+   else{
+    try {
+      await Postmodel.create({
+      userid:req.user.id,
+      description:desc,
+     }) 
+     return res.status(200).json({
+        success:true,
+
+      })
+    } catch (error) {
+      res.status(404).send(error.message)
+    }
+      
+   }
+} 
+
+
+
+
 
 // update a post
 //here the id which we are providing in params is of post id
@@ -37,13 +170,25 @@ exports.update = async (req,res)=>{
 
 }
 // delete a post
+//params id is of post you wanna delete
 exports.deletePost = async (req,res)=>{
+  
+  if(!req.params.id){
+    return res.status(404).send("post id not present in params")
+  }
 
- const post = await Postmodel.findById(req.params.id);
+  const _id = req.params.id
+ const post = await Postmodel.findById(_id);
+ 
  if(!post){
    return res.status(404).send("post not found")
  }
+ 
  try {
+  
+  if(post.img === null){
+    await cloudinary.v2.uploader.destroy(post.img.id)
+  }
    await post.deleteOne(); 
    res.status(200).send("Post has been deleted")
  } catch (error) {
